Float should map to warehouse FLOAT type
Float is a valid spark Datatype. There's no
reason to map floats to the larger double type

+++ elasticgraph-warehouse/spec/unit/elastic_graph/warehouse/schema_definition/scalar_type_extension_spec.rb
@@ -30,7 +30,7 @@ module ElasticGraph
 
           # Test the scalar types directly
           expect(warehouse_column_type_for(results, "Int")).to eq("INT")
-          expect(warehouse_column_type_for(results, "Float")).to eq("DOUBLE")
+          expect(warehouse_column_type_for(results, "Float")).to eq("FLOAT")
           expect(warehouse_column_type_for(results, "Boolean")).to eq("BOOLEAN")
           expect(warehouse_column_type_for(results, "String")).to eq("STRING")
           expect(warehouse_column_type_for(results, "ID")).to eq("STRING")
